all working .. now finish dialog window, looks like 0.85 is good ratio between ~0.6 and ~1.8 (the extreme case)

+++ MangaPrinter.WpfGUI/Utils/BucketInfo.cs
@@ -9,6 +9,6 @@ namespace MangaPrinter.WpfGUI.Utils
     public class BucketInfo {
         public int index = 0;
         public double value = 0;
-        public float count = 0;
+        public int count = 0;
     };
 }
